requirements.txt
python-telegram-bot==20.6
psycopg2-binary==2.9.10
python-dotenv==1.1.1
